refactor: move Bedrock stream timeout helpers
